requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==3.2.20
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
future==0.18.3
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
packaging==23.1
phonenumbers==8.13.16
phonenumberslite==8.13.16
pkg-resources==0.0.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0
serializers==0.2.4
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0